+++ src/main/java/ee/carlrobert/llm/client/codegpt/request/CodeCompletionRequest.java
@@ -1,6 +1,7 @@
 package ee.carlrobert.llm.client.codegpt.request;
 
 import ee.carlrobert.llm.completion.CompletionRequest;
+import java.util.List;
 import java.util.UUID;
 
 public class CodeCompletionRequest implements CompletionRequest {
@@ -10,8 +11,8 @@ public class CodeCompletionRequest implements CompletionRequest {
   private final String suffix;
   private final String fileExtension;
   private final String fileContent;
-  private final String stagedDiff;
-  private final String unstagedDiff;
+  private final int cursorOffset;
+  private final List<String> stop;
   private final UUID sessionId;
   private final String platformVersion;
   private final String pluginVersion;
@@ -22,8 +23,8 @@ public class CodeCompletionRequest implements CompletionRequest {
     this.suffix = builder.suffix;
     this.fileExtension = builder.fileExtension;
     this.fileContent = builder.fileContent;
-    this.stagedDiff = builder.stagedDiff;
-    this.unstagedDiff = builder.unstagedDiff;
+    this.cursorOffset = builder.cursorOffset;
+    this.stop = builder.stop;
     this.sessionId = builder.sessionId;
     this.platformVersion = builder.platformVersion;
     this.pluginVersion = builder.pluginVersion;
@@ -49,12 +50,12 @@ public class CodeCompletionRequest implements CompletionRequest {
     return fileContent;
   }
 
-  public String getStagedDiff() {
-    return stagedDiff;
+  public int getCursorOffset() {
+    return cursorOffset;
   }
 
-  public String getUnstagedDiff() {
-    return unstagedDiff;
+  public List<String> getStop() {
+    return stop;
   }
 
   public UUID getSessionId() {
@@ -71,13 +72,13 @@ public class CodeCompletionRequest implements CompletionRequest {
 
   public static class Builder {
 
-    private String model = "gpt-3.5-turbo-instruct";
+    private String model = "codestral";
     private String prefix;
     private String suffix;
     private String fileExtension;
     private String fileContent;
-    private String stagedDiff;
-    private String unstagedDiff;
+    private int cursorOffset;
+    private List<String> stop;
     private UUID sessionId;
     private String platformVersion;
     private String pluginVersion;
@@ -107,13 +108,13 @@ public class CodeCompletionRequest implements CompletionRequest {
       return this;
     }
 
-    public Builder setStagedDiff(String stagedDiff) {
-      this.stagedDiff = stagedDiff;
+    public Builder setCursorOffset(int cursorOffset) {
+      this.cursorOffset = cursorOffset;
       return this;
     }
 
-    public Builder setUnstagedDiff(String unstagedDiff) {
-      this.unstagedDiff = unstagedDiff;
+    public Builder setStop(List<String> stop) {
+      this.stop = stop;
       return this;
     }